Separate files and fix lines
