+++ CardGame/CardGame/Game/CardGame.swift
@@ -8,30 +8,12 @@
 
 import Foundation
 
-enum GameMode: Int, CaseIterable {
-    case fiveStud = 1, sevenStud
-    
-    var numberOfCards: Int {
-        switch self {
-        case .fiveStud:
-            return 5
-        case .sevenStud:
-            return 7
-        }
-    }
-}
-
-enum GameError: String, Error {
-    case noPlayer = "1명 이상의 플레이어가 필요합니다."
-    case noCard = "다음 게임을 진행하기에 카드가 부족합니다."
-}
-
 class CardGame {
     private let gameMode: GameMode
     private let numberOfPlayers: Int
     private var dealer: Dealer
     private var players: [GamePlayer] = []
-    
+
     init(gameMode: GameMode, numberOfPlayers: Int) throws {
         self.gameMode = gameMode
         guard numberOfPlayers > 0 else { throw GameError.noPlayer }
@@ -39,7 +21,7 @@ class CardGame {
         self.dealer = Dealer(with: CardDeck())
         setPlayers()
     }
-    
+
     private func setPlayers() {
         for number in 1...numberOfPlayers {
             let player = Player(number: number)
@@ -53,7 +35,7 @@ class CardGame {
             player.resetCards()
         }
     }
-    
+
     private func deal() throws {
         for _ in 1...gameMode.numberOfCards {
             for player in players {
