+++ CardGame/CardGame/Game/GameError.swift
@@ -1,0 +1,14 @@
+//
+//  GameError.swift
+//  CardGame
+//
+//  Created by 윤지영 on 16/11/2018.
+//  Copyright © 2018 JK. All rights reserved.
+//
+
+import Foundation
+
+enum GameError: Error {
+    case noPlayer
+    case noCard
+}
